requirements.txt
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
python-dotenv==1.1.0
psycopg2==2.9.10
pillow==11.2.1
phonenumbers==9.0.10
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.32.4